feat: 즐겨찾기 삭제 권한 체크

+++ src/main/java/nextstep/subway/favorite/application/FavoriteService.java
@@ -1,5 +1,6 @@
 package nextstep.subway.favorite.application;
 
+import nextstep.subway.auth.application.AuthorizationException;
 import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.favorite.domain.Favorite;
 import nextstep.subway.favorite.dto.FavoriteRequest;
@@ -48,6 +49,9 @@ public class FavoriteService {
 
     public void deleteFavorite(LoginMember loginMember, Long favoriteId) {
         Favorite favorite = favoriteRepository.findById(favoriteId).orElseThrow(RuntimeException::new);
+        if (!favorite.checkCreatedByMemberId(loginMember.getId())) {
+            throw new AuthorizationException("삭제 권한이 없습니다.");
+        }
         favoriteRepository.delete(favorite);
     }
 }
